New (fixed) patch set of metadata_adder discussion

+++ include/erasurecode/erasurecode_backend.h
@@ -109,7 +109,7 @@ struct ec_backend_common {
     char                        soversion[MAX_LEN]; /* EC backend shared library version */
 
     struct ec_backend_op_stubs  *ops;               /* EC backend stubs */
-    int                         metadata_adder;     /* EC backend custom metadata adder -
+    size_t                      metadata_adder;     /* EC backend custom metadata adder -
                                                      * metadata_adder bytes are added to
                                                      * the fragment size when allocating
                                                      * data/parity fragment buffers */
